requirements.txt
streamlit
pillow
opencv-python-headless
ultralytics
torch
pandas
